feat: NavFooter 구현 및 모든 페이지에 재사용
