+++ src/pages/OrderSuccessPage/OrderSuccessPage.tsx
@@ -1,33 +1,14 @@
 /** @jsxImportSource @emotion/react */
 
-import { css } from '@emotion/react';
 import { useSelectedCartItemsContext } from '../../features/cart/context/useSelectedCartItemsContext';
-import Button from '../../shared/ui/Button';
 import Navbar from '../../shared/ui/Navbar';
 import * as S from './OrderSuccessPage.styles';
 import { ROUTES } from '../../shared/constants/routeConstants';
 import { useNavigate } from 'react-router';
-
-const ButtonCSS = css`
-  width: 100%;
-  height: 100%;
-  background-color: #000;
-  color: #fff;
-  font-size: 16px;
-  font-weight: 700;
-  transition: background-color 0.2s ease;
-  &:hover {
-    background-color: #333;
-  }
-
-  &:disabled {
-    background-color: #bebebe;
-    cursor: not-allowed;
-  }
-`;
+import NavFooter from '../../shared/ui/NavFooter';
 
 export default function OrderSuccessPage() {
-  const { SelectedCartItems, cartTypeQuantity, totalQuantity, totalPurchasePrice } = useSelectedCartItemsContext();
+  const { cartTypeQuantity, totalQuantity, totalPurchasePrice } = useSelectedCartItemsContext();
 
   const navigate = useNavigate();
   const handleClick = () => {
@@ -46,9 +27,7 @@ export default function OrderSuccessPage() {
         <S.OrderSuccessTotalPurchasePriceLabel>총 결제 금액</S.OrderSuccessTotalPurchasePriceLabel>
         <S.OrderSuccessTotalPurchasePrice>{totalPurchasePrice.toLocaleString()}원</S.OrderSuccessTotalPurchasePrice>
       </S.OrderSuccessSection>
-      <S.OrderSuccessFooterContainer>
-        <Button onClick={handleClick} title="결제하기" css={ButtonCSS} disabled={SelectedCartItems.length === 0} />
-      </S.OrderSuccessFooterContainer>
+      <NavFooter title="장바구니로 돌아가기" onClick={handleClick} />
     </S.OrderSuccessContainer>
   );
 }
